pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zrsimacd"
version = "0.1.0"
description = "RSI/MACD toolkit & Tinkoff candles fetcher"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "<YOUR NAME>"}]
dependencies = [
    "pandas>=2.0",
    "pandas-ta>=0.3",
    "tinkoff",
    "python-dotenv>=1.0",
]

[project.scripts]
zrsimacd = "zrsimacd.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["zrsimacd"]
include-package-data = true